# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Arunayan Dairy API
  version: "1.0.0"
  description: Initial API contracts for Auth, Products, and Orders.
servers:
  - url: https://api.example.com
    description: Production
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
paths:
  /api/auth/signup:
    post:
      summary: Sign up a new user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthSignupRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/auth/login:
    post:
      summary: Login and obtain tokens
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/auth/refresh:
    post:
      summary: Refresh access token
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRefreshRequest'
      responses:
        '200':
          description: Tokens refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/products:
    get:
      summary: List products
      tags: [Products]
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: sort
          schema:
            type: string
            enum: [name.asc, name.desc, price.asc, price.desc, createdAt.desc]
            default: createdAt.desc
        - in: query
          name: minPrice
          schema:
            type: number
        - in: query
          name: maxPrice
          schema:
            type: number
        - in: query
          name: inStock
          schema:
            type: boolean
      responses:
        '200':
          description: Products page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductsPage'
    post:
      summary: Create a product
      tags: [Products]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/products/{id}:
    put:
      summary: Update a product
      tags: [Products]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdate'
      responses:
        '200':
          description: Product updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/orders:
    post:
      summary: Create an order
      tags: [Orders]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateRequest'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/orders/user/{userId}:
    get:
      summary: List orders for a user
      tags: [Orders]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnprocessableEntity:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
      required: [code, message]
    AuthSignupRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
      required: [name, email, password]
    AuthLoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
      required: [email, password]
    AuthRefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
      required: [refreshToken]
    AuthTokens:
      type: object
      properties:
        accessToken:
          type: string
        expiresIn:
          type: integer
          description: Seconds until expiry
        refreshToken:
          type: string
      required: [accessToken, expiresIn, refreshToken]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        tokens:
          $ref: '#/components/schemas/AuthTokens'
      required: [user, tokens]
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, email, role, createdAt, updatedAt]
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        price:
          type: number
        currency:
          type: string
          minLength: 3
          maxLength: 3
        sku:
          type: string
        stock:
          type: integer
          minimum: 0
        images:
          type: array
          items:
            type: string
            format: uri
        categoryId:
          type: string
          format: uuid
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, price, currency, stock, isActive, createdAt, updatedAt]
    ProductCreate:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 200
        description:
          type: string
        price:
          type: number
          minimum: 0
        currency:
          type: string
          minLength: 3
          maxLength: 3
        sku:
          type: string
        stock:
          type: integer
          minimum: 0
        images:
          type: array
          items:
            type: string
            format: uri
        categoryId:
          type: string
          format: uuid
        isActive:
          type: boolean
          default: true
      required: [name, price, currency, stock]
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 200
        description:
          type: string
        price:
          type: number
          minimum: 0
        currency:
          type: string
          minLength: 3
          maxLength: 3
        sku:
          type: string
        stock:
          type: integer
          minimum: 0
        images:
          type: array
          items:
            type: string
            format: uri
        categoryId:
          type: string
          format: uuid
        isActive:
          type: boolean
    OrderItem:
      type: object
      properties:
        productId:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
        unitPrice:
          type: number
          minimum: 0
        lineTotal:
          type: number
          minimum: 0
      required: [productId, quantity]
    Address:
      type: object
      properties:
        line1:
          type: string
        line2:
          type: string
        city:
          type: string
        state:
          type: string
        postalCode:
          type: string
        country:
          type: string
      required: [line1, city, state, postalCode, country]
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        totals:
          type: object
          properties:
            subtotal:
              type: number
            tax:
              type: number
            shipping:
              type: number
            total:
              type: number
          required: [subtotal, total]
        status:
          type: string
          enum: [pending, confirmed, shipped, delivered, cancelled]
        paymentStatus:
          type: string
          enum: [unpaid, paid, refunded]
        shippingAddress:
          $ref: '#/components/schemas/Address'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, userId, items, status, paymentStatus, createdAt]
    OrderCreateRequest:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: string
                format: uuid
              quantity:
                type: integer
                minimum: 1
            required: [productId, quantity]
          minItems: 1
        shippingAddress:
          $ref: '#/components/schemas/Address'
        paymentMethod:
          type: string
          enum: [cod, card, wallet]
        notes:
          type: string
      required: [items, shippingAddress, paymentMethod]
    PageInfo:
      type: object
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        totalItems:
          type: integer
        totalPages:
          type: integer
      required: [page, pageSize, totalItems, totalPages]
    ProductsPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        pageInfo:
          $ref: '#/components/schemas/PageInfo'
      required: [items, pageInfo]